Made install script more repeatable

+++ install.sh
@@ -61,22 +61,29 @@ if [ -z "${git_prog}" ]; then
 	sudo $PKGMGR git
 fi
 
-output "============== Installing VIM =========================="
-sudo $PKGMGR vim
+if which vim; then
+    output "============== Installing VIM =========================="
+    sudo $PKGMGR vim
+fi
 
-output "============== Installing Ag =========================="
 if [ "$OS" = "arch" ]; then
+    output "============== Installing Ag =========================="
     sudo $PKGMGR the_silver_searcher
 else
+    output "============== Installing Ag =========================="
     sudo $PKGMGR silversearcher-ag
 fi
 
-output "============== Installing ACPI utility =========================="
-sudo $PKGMGR acpi
+if which acpi; then
+    output "============== Installing ACPI utility =========================="
+    sudo $PKGMGR acpi
+fi
 
-output "=============== Cloning the repo ================="
-git clone https://github.com/turbshas/Setup_Scripts.git ~/.settings
-cd ~/.settings
+if [ ! -e ~/.settings ]; then
+    output "=============== Cloning the repo ================="
+    git clone https://github.com/turbshas/Setup_Scripts.git ~/.settings
+    cd ~/.settings
+fi
 
 #install zsh if not installed
 if [ -z "${zsh_prog}" ]; then
@@ -93,56 +100,60 @@ chmod +x oh.sh
 ./oh.sh
 rm -f ./oh.sh
 
+if [ -e ~/.oh-my-zsh ]; then
+    rm -rf ~/.oh-my-zsh
+fi
 #Add additional plugins for ZSH not found in oh-my-zsh
 git clone https://github.com/djui/alias-tips.git ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/alias-tips
 git clone https://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
 git clone https://github.com/marzocchi/zsh-notify.git ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/notify
 
 #setup softlinks
-output "=============== Setting up Soft Links ================="
-ln -s ~/.settings/vimrc ~/.vimrc
-mv ~/.zshrc ~/.zshrc.orig.ohmyzsh
+if [ ! -e ~/.vimrc ]; then
+    output "=============== Setting up Soft Links ================="
+    ln -s ~/.settings/vimrc ~/.vimrc
+fi
+if [ ! -e ~/.zshrc.orig.ohmyzsh ]; then
+    mv ~/.zshrc ~/.zshrc.orig.ohmyzsh
+fi
 echo "\"Add your custom vim settings to this file" > ~/.myvimrc
 echo "#Add your custom zsh settings to this file" > ~/.myzshrc
-ln -s ~/.settings/zshrc ~/.zshrc
-
-output "=============== Setting up cscope maps ====================="
-if [ ! -d ~/.vim ]; then
-	mkdir ~/.vim #if ~/.vim dir doesn't exist, create it
+if [ ~ -e ~/.zshrc ]; then
+    ln -s ~/.settings/zshrc ~/.zshrc
 fi
-if [ ! -d ~/.vim/plugin ]; then
-	mkdir ~/.vim/plugin #if plugin dir doesn't exist, create it
-fi
-cd ~/.vim/plugin
-curl -O http://cscope.sourceforge.net/cscope_maps.vim
-cd - # go back to the previous directory
 
-output "=============== Downloading badwolf theme ====================="
-if [ ! -d ${HOME}/.vim/colors ]; then
-    mkdir ${HOME}/.vim/colors
+if [ ! -e ~/.vim/plugin/cscope_maps.vim ]; then
+    output "=============== Setting up cscope maps ====================="
+	mkdir -p ~/.vim/plugin
+    cd ~/.vim/plugin
+    curl -O http://cscope.sourceforge.net/cscope_maps.vim
+    cd - # go back to the previous directory
 fi
-cd ${HOME}/.vim/colors
-curl -O https://raw.githubusercontent.com/sjl/badwolf/master/colors/badwolf.vim
-cd -
 
-output "=============== Setting up Makefile ftplugin ==============="
-if [ ! -d ~/.vim/ftplugin ]; then
-	mkdir ~/.vim/ftplugin
+if [ ! -e ~/.vim/colors/badwolf.vim ]; then
+    output "=============== Downloading badwolf theme ====================="
+    mkdir -p ${HOME}/.vim/colors
+    cd ${HOME}/.vim/colors
+    curl -O https://raw.githubusercontent.com/sjl/badwolf/master/colors/badwolf.vim
+    cd -
 fi
-cd ~/.vim/ftplugin
-#do not expand tabs in a makefile
-echo "setlocal noexpandtab" > make.vim
-cd -
 
-output "=============== Installing vim-gtk to get global clipboard support ==============="
+if [ ! -e ~/.vim/ftplugin/make.vim ]; then
+    output "=============== Setting up Makefile ftplugin ==============="
+	mkdir -p ~/.vim/ftplugin
+    cd ~/.vim/ftplugin
+    #do not expand tabs in a makefile
+    echo "setlocal noexpandtab" > make.vim
+    cd -
+fi
+
 if [ "$OS" = "ubuntu" ]; then
+    output "=============== Installing vim-gtk to get global clipboard support ==============="
     sudo $PKGMGR vim-gtk
 fi
 
 output "=============== Setup Go directories ==============="
-if [ ! -d ~/.go-dirs ]; then
-    mkdir ~/.go-dirs
-fi
+mkdir -p ~/.go-dirs
 
 #open vim once to install Plug
 vim +qall
@@ -150,9 +161,11 @@ vim +qall
 vim "+PlugInstall --sync" +qall
 
 output "=============== Install custom font ==============="
-mkdir /tmp/powerline
+mkdir -p /tmp/powerline
 cd /tmp/powerline
-git clone https://github.com/powerline/fonts.git
+if [ ! -e fonts ]; then
+    git clone https://github.com/powerline/fonts.git
+fi
 cd fonts
 #install all fonts
 ./install.sh
@@ -179,9 +192,11 @@ gsettings set org.gnome.Terminal.Legacy.Profile:/org/gnome/terminal/legacy/profi
 output "=============== Setting up Terminator ==============="
 ~/.settings/terminator/terminator.sh
 
+if [ ! -e ~/.oh-my-zsh/custom/themes/spaceship-prompt ]; then
 output "=============== Install custom zsh theme ==============="
-git clone https://github.com/denysdovhan/spaceship-prompt.git "${HOME}/.oh-my-zsh/custom/themes/spaceship-prompt"
-ln -s "${HOME}/.oh-my-zsh/custom/themes/spaceship-prompt/spaceship.zsh-theme" "${HOME}/.oh-my-zsh/themes/spaceship.zsh-theme"  
+    git clone https://github.com/denysdovhan/spaceship-prompt.git "${HOME}/.oh-my-zsh/custom/themes/spaceship-prompt"
+    ln -s "${HOME}/.oh-my-zsh/custom/themes/spaceship-prompt/spaceship.zsh-theme" "${HOME}/.oh-my-zsh/themes/spaceship.zsh-theme"
+fi
 
 output "=============== vimrc and zshrc files are now in your home folder ================="
 output "=============== Add custom vim settings to .myvimrc and zsh settings to .myzshrc files ============"
